movimento-api/src/main/java/br/com/jdsb/movimentoapi/controller/exceptions/ResourceExceptionHandler.java
package br.com.jdsb.movimentoapi.controller.exceptions;


import br.com.jdsb.movimentoapi.service.exception.MovimentoNaoEncontradoException;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

import javax.servlet.http.HttpServletRequest;
import java.time.LocalDateTime;

@ControllerAdvice
public class ResourceExceptionHandler {
    @ExceptionHandler(MovimentoNaoEncontradoException.class)
    public ResponseEntity<StandardError> movimentoNaoEncontrado(MovimentoNaoEncontradoException ex, HttpServletRequest request){
        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(
                new StandardError(LocalDateTime.now(),
                        ex.getMessage(),
                        HttpStatus.NOT_FOUND.value(),
                        request.getRequestURI())
        );
    }



}
