movimento-api/src/test/java/br/com/jdsb/movimentoapi/controller/exceptions/ResourceExceptionHandlerTest.java
package br.com.jdsb.movimentoapi.controller.exceptions;

import br.com.jdsb.movimentoapi.service.exception.MovimentoNaoEncontradoException;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.MockitoAnnotations;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.mock.web.MockHttpServletRequest;

import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
class ResourceExceptionHandlerTest {

    @InjectMocks
    private ResourceExceptionHandler exceptionHandler;

    public static final String MOVIMENTO_NAO_ENCONTRADO = "Movimento Não Encontrado";

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    void movimentoNaoEncontrado() {
        ResponseEntity<StandardError> response = exceptionHandler
                .movimentoNaoEncontrado(
                        new MovimentoNaoEncontradoException(MOVIMENTO_NAO_ENCONTRADO),
                        new MockHttpServletRequest());

        assertNotNull(response);
        assertNotNull(response.getBody());
        assertEquals(HttpStatus.NOT_FOUND, response.getStatusCode());
        assertEquals(ResponseEntity.class, response.getClass());
        assertEquals(StandardError.class, response.getBody().getClass());
        assertEquals(MOVIMENTO_NAO_ENCONTRADO, response.getBody().getError());
        assertEquals(404, response.getBody().getStatus());
    }
}